+++ ElvUI_MerathilisUI/Core/Changelog/6.78.lua
@@ -1,8 +1,14 @@
 local MER = unpack(ElvUI_MerathilisUI)
 
 MER.Changelog[678] = {
-	RELEASE_DATE = "TBD",
+	RELEASE_DATE = "24.10.2025",
 	FIXES = {},
-	NEW = {},
-	IMPROVEMENTS = {},
+	NEW = {
+		"[Misc]: Added a quick keystone feature",
+		"[Skins]: Added a skin for the CatalogShop",
+	},
+	IMPROVEMENTS = {
+		"[Skins]: Updated a tons of skins",
+		"[Quest]: Split the QuestProgress in an own module",
+	},
 }